% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{algebra}
\alias{algebra}
\title{Algebra for Predicted Marginal Group Means}
\usage{
algebra(dm)
}
\arguments{
\item{dm}{An object of class \code{designmatrix}.}
}
\value{
A data frame giving the combinations of grouping
  variables and then the sums of the regression
  coefficients needed to predict the marginal group means
  based on the specified design matrix. The final row
  (separated by '-') shows the sums over the regression
  coefficients for the predicted overal average.
}
\description{
Extracts a table displaying how the predicted marginal
group means would be calculated based on the
specified design matrix and a set of regression
coefficients. Also displays the algebra for the
overall group mean averaging over all groups.
}
\examples{
# Example with R data set 'PlantGrowth'
dm = designmatrix( PlantGrowth, list( 'weight', 'group' ) )
df = algebra( dm )

}
