% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_col.R
\name{design_col}
\alias{design_col}
\alias{design_col<-}
\title{Extract/Change Column Names for a Design Matrix}
\usage{
design_col(x)

design_col(x) <- value
}
\arguments{
\item{x}{An object of class \code{designmatrix}.}
}
\value{
The column names of the design matrix
  stored internally.
}
\description{
Extracts the column names for the design matrix
stored in a \code{designmatrix} object. Column
names can also be changed and updated in the
object.
}
\examples{
# Example with R data set 'PlantGrowth'
dm = designmatrix( PlantGrowth, list( 'weight', 'group' ) )
design_col( dm )
# Change names
design_col( dm ) = c( 'Control', 'Treat.1', 'Treat.2' )
print( dm )

}
