% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designmatrix-package.R
\docType{package}
\name{designmatrix-package}
\alias{designmatrix-package}
\title{Interfaces for Specifying Design Matrices}
\description{
The \pkg{designmatrix} package provides a set of
functions to streamline creation of design matrices
for statistical analyses (e.g., regression models
with planned contrasts). A complete  list of the
functions available in the package can be obtained
via the command \code{ls(pos = "package:designmatrix")}.
}
