% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designmatrix.R
\name{designmatrix}
\alias{designmatrix}
\alias{is.designmatrix}
\alias{print.designmatrix}
\alias{summary.designmatrix}
\alias{plot.designmatrix}
\alias{subset.designmatrix}
\alias{subset<-}
\title{Initialize and Update Design Matrices}
\usage{
designmatrix(df = NULL, select = NULL, dm = NULL, digits = 2)

is.designmatrix(x)

\method{print}{designmatrix}(x)

\method{summary}{designmatrix}(x)

\method{plot}{designmatrix}(x, intercept = T, exclude_effects = NULL,
  error_bars = F, average = T, ...)

\method{subset}{designmatrix}(x, summary = T)

subset(x, update = F) <- value
}
\arguments{
\item{df}{A data frame with the observations for
the dependent variable and associated grouping variables.}

\item{select}{A list with...
\enumerate{
  \item The column name for the dependent variable.
  \item The set of column names for the grouping variables.
}
If only the column name for the dependent variable is provided,
the function assumes all remaining columns are the set of
grouping variables.}

\item{dm}{An object of class \code{designmatrix}. If provided,
the function will update the object based on the specifications
of the \code{summary_matrix} element.}

\item{digits}{The number of digits to round to when
computing the group means.}
}
\value{
An object of class \code{designmatrix}. The \code{print}
  method displays the combination of levels and associated
  rows of the design matrix. The \code{summary} method
  displays the combination of levels and their associated
  descriptive statistics (mean, standard deviation,
  standard error of the mean, and sample size). The method
  \code{subset} can be used to access the summary matrix
  (or the full design matrix if \code{summary} is \code{TRUE}),
  or to replace elements of the summary matrix.

  The \code{plot} method generates a simple plot of the group
  means (filled circles). If elements of the design matrix are
  non-zero, the method fits a simple linear model to the data
  using the specified design matrix and shows the resulting
  estimates for the group means (empty circles). If the option
  \code{intercept} is set to \code{TRUE} (the default), the method
  assumes the intercept term is defined in the user-submitted
  design matrix. The option \code{error_bars}, if \code{TRUE},
  adds approximate 95% uncertainty intervals with boundaries
  equal to +/- 2 standard errors relative to the observed means.
  The option \code{exclude_effects} takes a string character
  with the column names for the design matrix. The method
  then displays what the estimated means would be were these
  columns excluded from from the design matrix (empty blue circles).
  Additional plotting options for \code{\link[graphics]{plot}},
  excluding \code{xlab}, \code{ylab}, \code{xaxt}, \code{pch},
  and \code{bty}, can be supplied.
}
\description{
Either 1) initialize a design matrix given a data frame and
a list specifying the dependent variable and associated
grouping variables, or 2) update a design matrix after
modification of of the element summarizing the design matrix
over all possible combinations of levels for the grouping
variables.
}
\examples{
# Example with R data set 'PlantGrowth'
dm = designmatrix( PlantGrowth, list( 'weight', 'group' ) )
# Summarize group means
summary( dm )

# Specify design matrix
# Main effect of treatment 1
subset( dm )[1:2,2] = c(-1,1)
# Main effect of treatment 2
subset( dm )[c(1,3),3] = c(-1,1)

# Update summaries and full design matrix
dm = designmatrix( dm )

# Example of methods
print( dm )
plot( dm, intercept = T, error_bars = T, exclude_effects = c( 'X2', 'X3' ) )

# Example analysis
dtba = getdata( dm )
lmf = lm( DV ~ -1 + ., data = dtba )

}
