% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Combined Data and Specified Design Matrices}
\usage{
getdata(dm)
}
\arguments{
\item{dm}{An object of class \code{designmatrix}.}
}
\value{
A data frame with a column for the dependent variable
  (labeled \code{DV}), and a remaining set of columns for the
  full design matrix extracted from the \code{designmatrix}
  object.
}
\description{
Extracts the dependent variable and full design
matrix from an object of class \code{designmatrix} and
combined them into a single data frame.
}
\examples{
# Example with R data set 'PlantGrowth'
dm = designmatrix( PlantGrowth, list( 'weight', 'group' ) )
df = getdata( dm )

}
