% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction}
\alias{interaction}
\title{Forthcoming}
\usage{
interaction(dm, V1, V2, start = NULL)
}
\arguments{
\item{dm}{Forthcoming.}

\item{V1}{Forthcoming.}

\item{V2}{Forthcoming.}

\item{start}{Forthcoming.}
}
\value{
Forthcoming.
}
\description{
Forthcoming
}
\examples{
# Example with R data set 'PlantGrowth'
dm = designmatrix( ToothGrowth, 'len' )

# Implement effect coding
subset( dm, T ) = coding( dm, type = 'EC', variable = 'supp', start = 2 )
subset( dm, T ) = coding( dm, type = 'EC', variable = 'dose', start = 3 )
subset( dm, T ) = interaction( dm, 'X2', c('X3','X4'), start = 5 )
print( dm )

}
