% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/designMatrix.R
\docType{package}
\name{designMatrix}
\alias{designMatrix}
\alias{designMatrix-package}
\title{designmatrix: A package for creating design matrices with dates.}
\usage{
designMatrix(x = NULL, weekdays = NULL, months = NULL, holidays = NULL,
  years = NULL, interactions = NULL, dayOfMonth = NULL,
  daysInMonth = NULL, leapYear = FALSE, quarters = NULL, weeks = NULL,
  weekend = FALSE, weekday = FALSE, removeUnusedLevels = FALSE,
  returnType = "numeric")
}
\arguments{
\item{x}{A vector of dates.}

\item{weekdays}{A character vector of the weekdays to include.}

\item{months}{A numeric vector of the months to include.}

\item{holidays}{TODO (A character vector of the holidays to include.)}

\item{years}{A numeric vector of the years to include.}

\item{interactions}{TODO (A character vector of the weekdays to include.)}

\item{dayOfMonth}{TODO A numeric vector of the day of the month to include.}

\item{daysInMonth}{TODO A numeric vector of the days in the month to include.}

\item{leapYear}{TODO A logical that includes to include a leap year indicator.}

\item{quarters}{TODO A character vector of the quarters to include.}

\item{weeks}{TODOA numeric vector to include the week number of the year.}

\item{weekend}{A logical that indicates if a weekend (Saturday or Sunday) indicator should be included}

\item{weekday}{A logical that indicates if a weekday (Monday through Friday) indicator should be included}

\item{removeUnusedLevels}{A TODO logical that indicates if colinear terms of the return matrix should be removed by dropping columns with zero variance.}

\item{returnType}{TODO A character vector indicating the data type for the return matrix. Acceptable values are "integer", "factor", "numeric", and "logical".}
}
\value{
The design matrix for the input dates.
}
\description{
The designmatrix package creates deisgn/model matrices for dates.
The workhorse designMatrix function supports many differetn terms
(e.g. weekday, month, leap years, holidays, etc). Some of these
are in development and as yet not implemented.
}
\section{designmatrix functions}{

designMatrix
}
\examples{
alldates <- seq(from = as.Date("2000-01-01"), to = as.Date("2005-10-01"), by = "days")

designMatrix(x = alldates, weekdays = "all", months = "January")
designMatrix(x = alldates, weekdays = c("Wednesday", "Friday"), quarters = "all")
designMatrix(x = alldates, months = "all", years = c(2002, 2004))
}

